requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.7.0
dj-database-url==2.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.6.0   # ✅ add this
